+++ inst/templates/ad_adae.R
@@ -4,7 +4,7 @@
 #
 # Input: ae, adsl, ex_single
 library(admiral)
-library(admiral.test) # Contains example datasets from the CDISC pilot project
+library(pharmaversesdtm) # Contains example datasets from the CDISC pilot project
 library(dplyr)
 library(lubridate)
 
@@ -14,12 +14,11 @@ library(lubridate)
 # as needed and assign to the variables below.
 # For illustration purposes read in admiral test data
 
-data("admiral_ae")
+data("ae")
 data("admiral_adsl")
 data("ex_single")
 
 adsl <- admiral_adsl
-ae <- admiral_ae
 suppae <- admiral_suppae
 
 # When SAS datasets are imported into R using haven::read_sas(), missing