+++ inst/templates/ad_adex.R
@@ -6,7 +6,7 @@
 #
 
 library(admiral)
-library(admiral.test) # Contains example datasets from the CDISC pilot project
+library(pharmaversesdtm) # Contains example datasets from the CDISC pilot project
 library(dplyr)
 library(lubridate)
 library(stringr)
@@ -19,7 +19,6 @@ data("admiral_adsl")
 data("admiral_ex")
 
 adsl <- admiral_adsl
-ex <- admiral_ex
 
 # When SAS datasets are imported into R using haven::read_sas(), missing
 # character values from SAS appear as "" characters in R, instead of appearing